PR comments + fix tests + move files to telemetry folder
