+++ x-pack/plugins/security_solution/public/common/components/control_columns/row_action/index.tsx
@@ -192,10 +192,10 @@ const RowActionComponent = ({
       },
     });
     telemetry.reportOpenNoteInExpandableFlyoutClicked({
-      location: timelineId,
+      location: tableId,
     });
     telemetry.reportDetailsFlyoutOpened({
-      location: timelineId,
+      location: tableId,
       panel: 'left',
     });
   }, [eventId, indexName, openFlyout, tableId, telemetry]);
