+++ x-pack/plugins/security_solution/public/flyout/document_details/left/components/add_note.tsx
@@ -113,7 +113,7 @@ export const AddNote = memo(({ eventId }: AddNewNoteProps) => {
       })
     );
     telemetry.reportAddNoteFromExpandableFlyoutClicked({
-      isRelatedToATimeline: checked && activeTimeline?.savedObjectId,
+      isRelatedToATimeline: checked && activeTimeline?.savedObjectId !== null,
     });
     setEditorValue('');
   }, [activeTimeline?.savedObjectId, checked, dispatch, editorValue, eventId, telemetry]);
